🎨 将检查用户关注功能从ImService移到FeedService
将检查用户是否关注的功能从ImService迁移到FeedService中，并完善了参数验证和错误处理逻辑，使功能归属更加合理
